+++ pyuarm/__init__.py
@@ -3,6 +3,5 @@ if sys.version > '3':
     PY3 = True
 else:
     PY3 = False
-from .uarm import UArm, get_uarm
+from .uarm import UArm, get_uarm, get_default_logger
 from .version import __version__
-from .util import set_debug